Move Toolbox into own widget for more control of sizing
