+++ app/src/mainwindow2.h
@@ -31,7 +31,7 @@ class ColorPaletteWidget;
 class OnionSkinWidget;
 class ToolOptionWidget;
 class TimeLine;
-class ToolBoxWidget;
+class ToolBoxDockWidget;
 class PreferencesDialog;
 class PreviewWidget;
 class ColorBox;
@@ -157,7 +157,7 @@ private:
     ColorBox*             mColorBox = nullptr;
     ColorPaletteWidget*   mColorPalette = nullptr;
     ToolOptionWidget*     mToolOptions = nullptr;
-    ToolBoxWidget*        mToolBox = nullptr;
+    ToolBoxDockWidget*        mToolBox = nullptr;
     RecentFileMenu*       mRecentFileMenu = nullptr;
     PreferencesDialog*    mPrefDialog = nullptr;
     //PreviewWidget*      mPreview = nullptr;
